Improve metadata widget
Each page in the metadata widget is now a tab, I implemented this according to the
designs in https://github.com/okfn/opendataeditor/issues/670.

Each tab contains the Resource metadata with its pages and the
form that corresponds to each page, as tabs they're positioned
side by side just like in the design and this makes it easier to
view each page.

Signed-off-by: Chihurumnaya Ibiam <ibiamchihurumnaya@gmail.com>