Improve metadata panel style
I've improved the style of metadata panel by directly applying
stylesheets to some of the widgets and also using a stylesheet
file to read some styles for other widgets.

The aim in the end is to keep the changes to widget in a stylesheet
so it's easier to manage.

I also set a RowWrapPolicy for the forms, this makes the form look
more like the initial design, although QFormLayout lays out its children
in a two-column form, with the labels in the left column and the
field widgets in the right column.
This makes it a bit difficult to completely follow the design as it has
two field widgets in each row, so we'll have to rethink this if we
want to stick to the design.

I thought about achieving the same thing using a grid, just haven't
tried it yet.

Signed-off-by: Chihurumnaya Ibiam <ibiamchihurumnaya@gmail.com>